backend/requirements.txt
aniso8601==10.0.1
blinker==1.9.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.7.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
libpass==1.9.3
MarkupSafe==3.0.3
python-dotenv==1.2.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
